refactor(store): deleteCable + reconnectCable in CableSlice (#308 Stufe 4)
projectStore.ts: 2277 → 2173 LOC (-104 LOC).

CableSlice komplettiert: deleteCable und reconnectCable in cableSlice.ts
hinzugefuegt. Damit lebt die komplette Cable-Domain-Logik in einer Datei
(8 Actions, 270 LOC) — bis auf addCableFromMobile, das wegen seiner
Verwobenheit mit scheduleProjectAutosave + projectVersion-Increment
vorerst in projectStore.ts bleibt.

Gesamt #308 seit Start: 2589 → 2173 LOC (-16%).
